requirements.txt
mysql-connector-python==8.0.23
Pillow==8.1.2
prettytable==3.3.0
protobuf==3.15.6
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PyQtChart==5.15.4
PyQtChart-Qt5==5.15.2
reportlab==3.5.65
six==1.15.0
wcwidth==0.2.5


altgraph==0.17.3
appdirs==1.4.3
apturl==0.5.2
attrs==23.1.0
auto-py-to-exe==2.37.0
autopep8==1.6.0
bcrypt==3.1.7
bitarray==2.6.1
blinker==1.4
bottle==0.12.25
bottle-websocket==0.2.9
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2023.7.22
cffi==1.15.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
Eel==0.16.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fasteners==0.14.1
filelock==3.0.12
flake8==4.0.1
fonttools==4.36.0
fpdf==1.7.2
future==0.18.2
gevent==23.7.0
gevent-websocket==0.10.1
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
greenlet==2.0.2
h11==0.14.0
httplib2==0.14.0
iconify==0.0.103
idna==2.8
importlib-metadata==1.5.0
jsonschema==3.2.0
keyring==18.0.1
kids.cache==0.0.7
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libsass==0.21.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.3
mccabe==0.6.1
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
mypy==0.950
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
netifaces==0.10.4
numpy==1.23.2
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.1.2
outcome==1.2.0
packaging==21.3
pandas==2.0.3
paramiko==2.6.0
pdfreader==0.1.12
pdfrw2==0.5.0
pexpect==4.6.0
Pillow==8.1.2
prettytable==3.0.0
protobuf==3.15.6
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.16.0
pycups==1.9.73
PyDrive==1.3.1
pyflakes==2.4.0
PyGObject==3.36.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
PyQtChart==5.15.4
PyQtChart-Qt5==5.15.2
pyRFC3339==1.1
pyrsistent==0.15.5
PySide2==5.15.2.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
QT-PyQt-PySide-Custom-Widgets==0.6.3
qtsass==0.3.0
reportlab==3.5.65
requests==2.23.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
selenium==4.11.2
shiboken2==5.15.2.1
simplejson==3.16.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.12
systemd-python==234
texttable==1.6.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
tqdn==0.0.8
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.2.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==2.0.4
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.53.0
whichcraft==0.6.1
wsproto==1.2.0
xkit==0.0.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.0
